requirements.txt
Pillow==4.3.0
pystray==0.14.3
